requirements.txt
streamlit
plotly
streamlit-lottie
fpdf2
selenium
pillow
pdfkit
weasyprint